python/requirements.txt
# IDE requirements
autopep8
pylint
pycodestyle

# Requirements
Flask
psycopg2-binary
